Rename trellis-io-jena to trellis-jena
