+++ components/dropwizard/src/test/java/org/trellisldp/dropwizard/SimpleServiceBundler.java
@@ -31,8 +31,8 @@ import org.trellisldp.audit.DefaultAuditService;
 import org.trellisldp.constraint.LdpConstraintService;
 import org.trellisldp.file.FileBinaryService;
 import org.trellisldp.http.core.DefaultTimemapGenerator;
-import org.trellisldp.io.JenaIOService;
-import org.trellisldp.io.NoopProfileCache;
+import org.trellisldp.jena.JenaIOService;
+import org.trellisldp.jena.NoopProfileCache;
 import org.trellisldp.triplestore.TriplestoreResourceService;
 
 /**
