--- components/io-jena/src/main/resources/META-INF/services/org.trellisldp.api.IOService
@@ -1,1 +1,0 @@
-org.trellisldp.io.JenaIOService
