+++ components/jena/src/main/java/module-info.java
@@ -13,14 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trellisldp.io;
+module org.trellisldp.jena {
+    exports org.trellisldp.jena;
 
-import org.trellisldp.api.CacheService.TrellisProfileCache;
-import org.trellisldp.api.NoopCacheService;
+    requires transitive org.trellisldp.api;
+    requires transitive org.trellisldp.vocabulary;
 
-/**
- * A specialized no-op cache.
- */
-@TrellisProfileCache
-public class NoopProfileCache extends NoopCacheService<String, String> {
+    requires org.apache.commons.io;
+    requires org.apache.commons.rdf.api;
+    requires org.apache.jena.arq;
+    requires org.apache.jena.base;
+    requires org.apache.jena.commonsrdf;
+    requires org.apache.jena.core;
+    requires jakarta.enterprise.cdi.api;
+    requires jakarta.inject;
+    requires microprofile.config.api;
+    requires org.slf4j;
+
+    provides org.trellisldp.api.IOService
+        with org.trellisldp.jena.JenaIOService;
 }
