+++ components/jena/src/main/resources/META-INF/beans.xml
