+++ components/jena/src/main/resources/META-INF/services/org.trellisldp.api.IOService
@@ -1,0 +1,1 @@
+org.trellisldp.jena.JenaIOService
