+++ components/jena/src/test/java/org/trellisldp/jena/NoopProfileCacheTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trellisldp.io;
+package org.trellisldp.jena;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
