+++ components/jena/src/test/resources/testRdf.ttl
