+++ components/test/build.gradle
@@ -34,8 +34,8 @@ dependencies {
     implementation "org.hamcrest:hamcrest:$hamcrestVersion"
     implementation "org.slf4j:slf4j-api:$slf4jVersion"
     implementation project(':trellis-api')
+    implementation project(':trellis-jena')
     implementation project(':trellis-http')
-    implementation project(':trellis-io-jena')
     implementation project(':trellis-vocabulary')
 
     testImplementation "ch.qos.logback:logback-classic:$logbackVersion"
