+++ components/test/src/main/java/org/trellisldp/test/TestUtils.java
@@ -46,8 +46,8 @@ import org.apache.commons.rdf.api.RDFSyntax;
 import org.trellisldp.api.IOService;
 import org.trellisldp.api.NoopNamespaceService;
 import org.trellisldp.api.RDFFactory;
-import org.trellisldp.io.JenaIOService;
-import org.trellisldp.io.NoopProfileCache;
+import org.trellisldp.jena.JenaIOService;
+import org.trellisldp.jena.NoopProfileCache;
 import org.trellisldp.vocabulary.AS;
 import org.trellisldp.vocabulary.LDP;
 import org.trellisldp.vocabulary.PROV;
