+++ core/http/build.gradle
@@ -7,7 +7,7 @@ description = 'Trellis HTTP'
 
 ext {
     moduleName = 'org.trellisldp.http'
-    testModules = ['org.trellisldp.io', 'org.trellisldp.audit', 'org.trellisldp.constraint', 'smallrye.config',
+    testModules = ['org.trellisldp.jena', 'org.trellisldp.audit', 'org.trellisldp.constraint', 'smallrye.config',
         'com.fasterxml.jackson.databind', 'org.apache.jena.commonsrdf', 'com.fasterxml.jackson.core']
 }
 
@@ -40,8 +40,8 @@ dependencies {
     testImplementation "org.mockito:mockito-core:$mockitoVersion"
     testImplementation project(':trellis-audit')
     testImplementation project(':trellis-constraint')
+    testImplementation project(':trellis-jena')
     testImplementation project(':trellis-event-jackson')
-    testImplementation project(':trellis-io-jena')
 
     testRuntimeOnly "jakarta.activation:jakarta.activation-api:$activationApiVersion"
     testRuntimeOnly "ch.qos.logback:logback-classic:$logbackVersion"
