+++ platform/bom/build.gradle
@@ -23,8 +23,8 @@ dependencies {
     implementation project(':trellis-constraint')
     implementation project(':trellis-dropwizard')
     implementation project(':trellis-file')
-    implementation project(':trellis-io-jena')
     implementation project(':trellis-jdbc')
+    implementation project(':trellis-jena')
     implementation project(':trellis-namespace')
     implementation project(':trellis-rdfa')
     implementation project(':trellis-test')
