+++ platform/dropwizard/build.gradle
@@ -38,25 +38,25 @@ dependencies {
     implementation "org.apache.jena:jena-tdb2:$jenaVersion"
     implementation "org.slf4j:slf4j-api:$slf4jVersion"
 
-    implementation project(':trellis-constraint')
-    implementation project(':trellis-io-jena')
     implementation project(':trellis-api')
-    implementation project(':trellis-http')
     implementation project(':trellis-app')
-    implementation project(':trellis-cache')
-    implementation project(':trellis-dropwizard')
-    implementation project(':trellis-vocabulary')
-    implementation project(':trellis-file')
-    implementation project(':trellis-namespace')
     implementation project(':trellis-audit')
-    implementation project(':trellis-event-jackson')
-    implementation project(':trellis-webac')
-    implementation project(':trellis-triplestore')
-    implementation project(':trellis-jms')
-    implementation project(':trellis-rdfa')
-    implementation project(':trellis-kafka')
-    implementation project(':trellis-oauth')
     implementation project(':trellis-auth-basic')
+    implementation project(':trellis-cache')
+    implementation project(':trellis-constraint')
+    implementation project(':trellis-dropwizard')
+    implementation project(':trellis-event-jackson')
+    implementation project(':trellis-file')
+    implementation project(':trellis-http')
+    implementation project(':trellis-jena')
+    implementation project(':trellis-jms')
+    implementation project(':trellis-kafka')
+    implementation project(':trellis-namespace')
+    implementation project(':trellis-oauth')
+    implementation project(':trellis-rdfa')
+    implementation project(':trellis-triplestore')
+    implementation project(':trellis-vocabulary')
+    implementation project(':trellis-webac')
 
     implementation "io.dropwizard:dropwizard-core:$dropwizardVersion"
     implementation "io.dropwizard:dropwizard-metrics:$dropwizardVersion"
