+++ platform/karaf/src/main/resources/features.xml
@@ -108,7 +108,7 @@
     <bundle>mvn:org.trellisldp/trellis-constraint/${project.version}</bundle>
   </feature>
 
-  <feature name="trellis-io-jena" version="${project.version}">
+  <feature name="trellis-jena" version="${project.version}">
     <details>Installs the Jena-backed serialization service</details>
 
     <feature dependency="true">commons-rdf-jena</feature>
@@ -118,7 +118,7 @@
     <bundle dependency="true">mvn:org.apache.servicemix.bundles/org.apache.servicemix.bundles.javax-inject/1_2</bundle>
     <bundle dependency="true">mvn:org.eclipse.microprofile.config/microprofile-config-api/${microprofileConfigVersion}</bundle>
 
-    <bundle>mvn:org.trellisldp/trellis-io-jena/${project.version}</bundle>
+    <bundle>mvn:org.trellisldp/trellis-jena/${project.version}</bundle>
 
   </feature>
