+++ platform/osgi/build.gradle
@@ -61,8 +61,8 @@ dependencies {
     testImplementation project(':trellis-event-jsonb')
     testImplementation project(':trellis-file')
     testImplementation project(':trellis-http')
-    testImplementation project(':trellis-io-jena')
     testImplementation project(':trellis-jdbc')
+    testImplementation project(':trellis-jena')
     testImplementation project(':trellis-jms')
     testImplementation project(':trellis-kafka')
     testImplementation project(':trellis-namespace')
