+++ platform/quarkus/build.gradle
@@ -32,8 +32,8 @@ dependencies {
     implementation project(":trellis-event-jsonb")
     implementation project(":trellis-file")
     implementation project(":trellis-http")
+    implementation project(":trellis-jena")
     implementation project(":trellis-jwt")
-    implementation project(":trellis-io-jena")
     implementation project(":trellis-namespace")
     implementation project(":trellis-rdfa")
     implementation project(":trellis-reactive")
